# java/src/main/resources/openapi-chat.yaml
openapi: 3.0.3
info:
  title: Netdisk API - 聊天功能
  description: 网盘系统聊天功能接口文档
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: 本地开发环境
paths:
...
  /api/v2/chat/conversations:
    post:
      tags:
        - 会话管理
      summary: 创建会话
      description: 创建群聊或私聊会话
      operationId: createConversation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    get:
      tags:
        - 会话管理
      summary: 获取会话列表
      description: 获取当前用户的会话列表，支持分页和搜索
      operationId: listConversations
      security:
        - BearerAuth: []
      parameters:
        - name: keyword
          in: query
          description: 搜索关键字（会话名称）
          schema:
            type: string
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          description: 每页条数
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationListResponse'
  /api/v2/chat/conversations/{conversationId}:
    get:
      tags:
        - 会话管理
      summary: 获取会话详情
      description: 获取指定会话的详细信息，包括成员列表
      operationId: getConversation
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetailResponse'

    patch:
      tags:
        - 会话管理
      summary: 更新会话
      description: 修改会话名称或头像
      operationId: updateConversation
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConversationRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetailResponse'
  /api/v2/chat/conversations/{conversationId}/archive:
    post:
      tags:
        - 会话管理
      summary: 归档会话
      description: 将会话移入归档列表
      operationId: archiveConversation
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
      responses:
        '200':
          description: 归档成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArchiveResponse'
  /api/v2/chat/conversations/{conversationId}/unarchive:
    post:
      tags:
        - 会话管理
      summary: 取消归档
      description: 恢复已归档的会话
      operationId: unarchiveConversation
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
      responses:
        '200':
          description: 恢复成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnarchiveResponse'

  # ==================== 成员管理 ====================
  /api/v2/chat/conversations/{conversationId}/members:
    get:
      tags:
        - 成员管理
      summary: 获取成员列表
      description: 获取会话的成员列表
      operationId: listMembers
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberListResponse'

    post:
      tags:
        - 成员管理
      summary: 添加成员
      description: 向群聊添加成员
      operationId: addMembers
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMembersRequest'
      responses:
        '200':
          description: 添加成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddMembersResponse'
  /api/v2/chat/conversations/{conversationId}/members/{userId}:
    delete:
      tags:
        - 成员管理
      summary: 移除成员
      description: 从群聊中移除指定成员（仅群主/管理员可执行）
      operationId: removeMember
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
        - name: userId
          in: path
          required: true
          description: 被移除的用户UUID
          schema:
            type: string
      responses:
        '200':
          description: 移除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoveMemberResponse'
  /api/v2/chat/conversations/{conversationId}/leave:
    post:
      tags:
        - 成员管理
      summary: 主动退群
      description: 成员主动退出群聊（群主无法退群，需先转让或解散）
      operationId: leaveConversation
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
      responses:
        '200':
          description: 退群成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveResponse'
        '403':
          description: 群主无法退群
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v2/chat/conversations/{conversationId}/dissolve:
    post:
      tags:
        - 成员管理
      summary: 解散群聊
      description: 群主/管理员解散群聊
      operationId: dissolveConversation
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
      responses:
        '200':
          description: 解散成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DissolveResponse'
        '400':
          description: 私聊无法解散
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v2/chat/conversations/{conversationId}/transfer:
    post:
      tags:
        - 成员管理
      summary: 转让群主
      description: 群主将群主身份转让给其他成员
      operationId: transferOwnership
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferOwnershipRequest'
      responses:
        '200':
          description: 转让成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferOwnershipResponse'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v2/chat/conversations/{conversationId}/admins:
    post:
      tags:
        - 成员管理
      summary: 设置管理员
      description: 群主/管理员将成员设置为管理员
      operationId: setAdmin
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetAdminRequest'
      responses:
        '200':
          description: 设置成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetAdminResponse'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v2/chat/conversations/{conversationId}/admins/{userId}:
    delete:
      tags:
        - 成员管理
      summary: 取消管理员
      description: 群主/管理员取消成员的管理员身份
      operationId: removeAdmin
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
        - name: userId
          in: path
          required: true
          description: 被取消管理员的用户UUID
          schema:
            type: string
      responses:
        '200':
          description: 取消成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoveAdminResponse'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ==================== 免打扰 ====================
  /api/v2/chat/conversations/{conversationId}/mute:
    post:
      tags:
        - 免打扰
      summary: 设置免打扰
      description: 设置会话免打扰
      operationId: muteConversation
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MuteRequest'
      responses:
        '200':
          description: 设置成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MuteResponse'
  /api/v2/chat/conversations/{conversationId}/unmute:
    post:
      tags:
        - 免打扰
      summary: 取消免打扰
      description: 取消会话免打扰
      operationId: unmuteConversation
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
      responses:
        '200':
          description: 取消成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnmuteResponse'

  # ==================== 消息管理 ====================
  /api/v2/chat/conversations/{conversationId}/messages:
    post:
      tags:
        - 消息管理
      summary: 发送消息
      description: 在会话中发送消息
      operationId: sendMessage
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: 发送成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

    get:
      tags:
        - 消息管理
      summary: 获取消息历史
      description: 获取会话的消息历史（游标分页）
      operationId: listMessages
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
        - name: cursor
          in: query
          description: 游标（上一页最后一条消息的ID）
          schema:
            type: string
        - name: limit
          in: query
          description: 每页条数（1-100）
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListResponse'
  /api/v2/chat/messages/{messageId}:
    patch:
      tags:
        - 消息管理
      summary: 编辑消息
      description: 修改消息内容
      operationId: editMessage
      security:
        - BearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          description: 消息UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditMessageRequest'
      responses:
        '200':
          description: 编辑成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

    delete:
      tags:
        - 消息管理
      summary: 撤回消息
      description: 撤回发送的消息（仅支持2分钟内的消息撤回）
      operationId: recallMessage
      security:
        - BearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          description: 消息UUID
          schema:
            type: string
      responses:
        '200':
          description: 撤回成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecallResponse'
  /api/v2/chat/messages/{messageId}/read:
    post:
      tags:
        - 消息管理
      summary: 标记已读
      description: 将单条消息标记为已读
      operationId: markRead
      security:
        - BearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          description: 消息UUID
          schema:
            type: string
      responses:
        '200':
          description: 标记成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarkReadResponse'
  /api/v2/chat/messages/read-batch:
    post:
      tags:
        - 消息管理
      summary: 批量已读
      description: 异步批量标记消息为已读
      operationId: markReadBatch
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadBatchRequest'
      responses:
        '200':
          description: 提交成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadBatchResponse'

  # ==================== 未读数 ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          type: object
          description: 响应数据

    # ==================== 请求模型 ====================
    CreateConversationRequest:
      type: object
      required:
        - conversationType
        - memberUserIds
      properties:
        conversationType:
          type: string
          enum: [group, direct]
          description: 会话类型（group=群聊，direct=私聊）
          example: "group"
        spaceId:
          type: string
          description: 关联空间ID（可选）
          example: "space_xxx"
        name:
          type: string
          description: 群聊名称（最大128字符）
          example: "技术交流群"
        avatarUrl:
          type: string
          description: 群头像URL
          example: "https://xxx.com/avatar.jpg"
        memberUserIds:
          type: array
          description: 成员用户UUID列表（群聊上限500人，超出部分返回failedMembers）
          items:
            type: string
          example: ["user_uuid_1", "user_uuid_2"]
...
    ConversationResponse:
      type: object
      properties:
        id:
          type: string
          description: 会话UUID
        conversationType:
          type: string
        name:
          type: string
        avatarUrl:
          type: string
        status:
          type: string
        ownerUserId:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/MemberInfo'
...
    ArchiveResponse:
      type: object
      properties:
        conversationId:
          type: string
        archived:
          type: boolean
...